+++ demo/lib/TransformationControls/addCompatibilityControl.js
@@ -1,12 +1,18 @@
 /* eslint-disable require-jsdoc */
 
+const enableCompatibilitySupport = (iframeWindow, iframeDocument) => {
+  event.target.disabled = true
+  if (iframeWindow.pagelib.c1) {
+    // TODO: compatibility transform is not exposed in the abstraction layer
+  } else {
+    iframeWindow.pagelib.CompatibilityTransform.enableSupport(iframeDocument)
+  }
+}
+
 export default [
   'compatibility_checkbox',
   'Compatibility filter *',
   'checkbox',
   null,
-  (iframeWindow, iframeDocument) => {
-    event.target.disabled = true
-    iframeWindow.pagelib.CompatibilityTransform.enableSupport(iframeDocument)
-  }
+  enableCompatibilitySupport,
 ]